requirements.txt
numpy>=1.23
pandas>=2.0
matplotlib>=3.7
seaborn>=0.13
lifelines>=0.28
statsmodels>=0.14
pyyaml>=6.0
pytest>=7.0
types-PyYAML>=6.0
pytest-cov>=4.0
pandera>=0.18
